app/requirements.txt
langchain-text-splitters 
langchain-community
langchain-openai
python-dotenv
langchain-huggingface
langchain-chroma


